@@ -1,16 +1,14 @@
 package edu.eci.arsw.msgbroker;
 
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
+import edu.eci.arsw.collabpaint.CollabPaintApplication;
 
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class WebsocketsMsbrokerExampleApplicationTests {
+@SpringBootTest(classes = CollabPaintApplication.class)
+class WebsocketsMsbrokerExampleApplicationTests {
 
 	@Test
-	public void contextLoads() {
+	void contextLoads() {
 	}
 
 }
