@@ -36,9 +36,10 @@ var app = (function () {
         //subscribe to /topic/TOPICXX when connections succeed
         stompClient.connect({}, function (frame) {
             console.log('Connected: ' + frame);
-            stompClient.subscribe('/topic/TOPICXX', function (eventbody) {
-                
-                
+            stompClient.subscribe('/topic/newpoint', function (eventbody) {
+                var theObject = JSON.parse(eventbody.body);
+                alert("Nuevo punto recibido: X=" + theObject.x + ", Y=" + theObject.y);
+                addPointToCanvas(theObject);
             });
         });
 
@@ -60,7 +61,7 @@ var app = (function () {
             console.info("publishing point at "+pt);
             addPointToCanvas(pt);
 
-            //publicar el evento
+            stompClient.send("/topic/newpoint", {}, JSON.stringify(pt));
         },
 
         disconnect: function () {
